% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{corrected_words}
\alias{corrected_words}
\title{Available Datasets
A dataset of corrected words}
\format{
A tibble with corrected words.
}
\usage{
corrected_words
}
\description{
A collection of words corrected from user-generated text.
}
\keyword{datasets}
