% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load dataset}
\usage{
load_data(format = "csv")
}
\arguments{
\item{format}{The file format to load ("csv" or "rds")}
}
\value{
A data frame
}
\description{
Load dataset
}
\examples{
tidyr::pivot_longer(corrected_words_10137,
                    cols = everything(),  # Pivot all columns to long format
                                     names_to = "word", # Column name"
                                      values_to = "corrected_word") # Changed column name"
}
