% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reviews.R
\name{load_reviews}
\alias{load_reviews}
\title{Load Review Data from RDS}
\usage{
load_reviews()
}
\value{
A tibble containing review data.
}
\description{
This function loads the main dataset of corrected words or product reviews from
an RDS file within the package.
}
