% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_tr.R
\name{load_stopwords}
\alias{load_stopwords}
\title{Load stopwords dataset}
\usage{
load_stopwords(format = "csv")
}
\arguments{
\item{format}{The file format to load ("csv")}
}
\value{
A data frame (tibble)
}
\description{
Load stopwords dataset
}
