% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_match.R
\name{normalize_text}
\alias{normalize_text}
\title{Load dataset}
\usage{
normalize_text(text)
}
\arguments{
\item{format}{The file format to load ("csv")}
}
\value{
A data frame
}
\description{
Load dataset
}
\examples{
user_input <- "Göre"
match <- amatch("göre", stopwords_all, method = "lv", maxDist = 1)
normalized_input <- normalize_text(user_input)
}
