% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{A dataset of phrases}
\format{
A tbl_df with with 205 rows and 1 variable:
\describe{
\item{word}{ngrams.}
}
}
\usage{
phrases
}
\description{
Contains common negative-emotion phrases extracted from user reviews.
}
\keyword{datasets}
