% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\docType{data}
\name{reviews}
\alias{reviews}
\title{A dataset of reviews}
\format{
A tbl_df with with 260,308 rows and 3 variables:
\describe{
\item{rating}{Rating score, out of 5.}
\item{comment}{Comment text, in Turkish.}
\item{id}{Rating ID.}
}
}
\usage{
reviews
}
\description{
User reviews collected from an e-commerce site.
}
\keyword{datasets}
