% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews_test.R
\docType{data}
\name{reviews_test}
\alias{reviews_test}
\title{A test dataset}
\format{
A tbl_df with with 1,481 rows and 3 variables:
\describe{
\item{rating}{Rating score, out of 5.}
\item{text}{Comment text, in Turkish.}
\item{emotion}{n for negative, p for positive.}
\item{id}{Rating ID.}
}
}
\usage{
reviews_test
}
\description{
A test sample data used for testing analysis functions. It differs from reviews data.
The \code{text} column in this data frame is similar to the \code{comment} column in the \code{reviews}
data frame. Note that this data frame contains 170 \code{text}s that are in common, verbatim,
with comments in the \code{reviews} dataset. This is because some users made the same comments.
The \code{id} column shows that these are not the same observations, just similarly worded
comments from different reviews.
}
\keyword{datasets}
