% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoppingwords-package.R
\docType{package}
\name{shoppingwords-package}
\alias{shoppingwords}
\alias{shoppingwords-package}
\title{shoppingwords: It includes the collection of words from user reviews}
\description{
It includes the collection of words from user reviews that were captured from an online shopping platform.
}
\author{
\strong{Maintainer}: Betul Kan-Kilinc \email{bk234@duke.edu}

Other contributors:
\itemize{
  \item Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} [contributor]
  \item Colin Rundel \email{cr173@duke.edu} [contributor]
}

}
\keyword{internal}
