% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{stopwords_tr}
\alias{stopwords_tr}
\title{A dataset of Turkish stopwords}
\format{
A tibble with stopwords.
}
\usage{
stopwords_tr
}
\description{
A dataset of stopwords used in Turkish text analysis.
}
\keyword{datasets}
