% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{A test dataset}
\format{
A tibble.
}
\usage{
testdata
}
\description{
A test sample data used for testing analysis functions. It differs from reviews data.
}
\keyword{datasets}
