% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corrected_words.R
\name{load_corrected_words}
\alias{load_corrected_words}
\title{Load dataset
This function loads corrected words from reviews and from a CSV (json) file within the package.}
\usage{
load_corrected_words(format = "csv")
}
\arguments{
\item{format}{The file format to load ("csv", or "json")}
}
\value{
A data frame
}
\description{
Load dataset
This function loads corrected words from reviews and from a CSV (json) file within the package.
}
