% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_phrases.R
\name{load_phrases}
\alias{load_phrases}
\title{Load predefined phrases}
\usage{
load_phrases(format = "csv")
}
\arguments{
\item{format}{Character string specifying the format. Default is "csv".}
}
\value{
A tibble with predefined phrases.
}
\description{
This function loads predefined phrases from a CSV file within the package.
}
