% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stopwords.R
\name{load_stopwords}
\alias{load_stopwords}
\title{Load predefined stopwords list for Turkish shopping-related text analysis}
\usage{
load_stopwords(format = "csv")
}
\arguments{
\item{format}{Character string specifying the format. Default is "csv".}
}
\value{
A tibble containing predefined Turkish stopwords shopping based-special.
}
\description{
This function retrieves a predefined list of 92 stopwords from a CSV file
stored within the package. These stopwords are commonly used in Turkish
shopping-related text data
}
