% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_tr.R
\name{load_stopwords}
\alias{load_stopwords}
\title{Load predefined stopwords}
\usage{
load_stopwords(format = "csv")
}
\arguments{
\item{format}{Character string specifying the format. Default is "csv".}
}
\value{
A tibble with stopwords.
}
\description{
This function loads stopwords from a CSV file within the package.
}
