% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stopwords.R
\name{match_stopwords}
\alias{match_stopwords}
\title{Normalize Stopwords in Text}
\usage{
match_stopwords(df)
}
\arguments{
\item{text}{Character string representing the word to normalize.}
}
\value{
A list containing the matched word and its corresponding index, or \code{NULL} if no match is found.
}
\description{
This function checks whether a given word is present in a predefined stopwords list.
It first searches the package's internal stopwords dataset (\code{stopwords_tr}), and if
no match is found, it falls back to the broader \code{stopwords_iso} list.
}
\details{
The function converts text to a standard format by removing accents and special characters,
transforming it into basic Latin characters, and making all letters lowercase.
This ensures more accurate word comparisons.
}
