% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_text.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize stopwords
This function loads stopwords from a CSV file within the package.}
\usage{
normalize_text(text)
}
\arguments{
\item{text}{Input word to normalize}
}
\value{
Normalized version of the input word
}
\description{
Normalize stopwords
This function loads stopwords from a CSV file within the package.
}
