% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{phrases}
\alias{phrases}
\title{A dataset of phrases}
\format{
A tibble with phrases.
}
\usage{
phrases
}
\description{
Contains common negative-emotion phrases extracted from user reviews.
}
\keyword{datasets}
