% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{reviews}
\alias{reviews}
\title{A dataset of reviews}
\format{
A tibble containing text reviews and rating scores.
}
\usage{
reviews
}
\description{
User reviews collected from an e-commerce site.
}
\keyword{datasets}
