% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoppingwords-package.R
\docType{package}
\name{shoppingwords-package}
\alias{shoppingwords}
\alias{shoppingwords-package}
\title{shoppingwords: Text Processing Tools for Turkish E-Commerce Data}
\description{
Provides several datasets useful for processing and analysis of text in Turkish from an online shopping platform.
}
\author{
\strong{Maintainer}: Betul Kan-Kilinc \email{bk234@duke.edu}

Other contributors:
\itemize{
  \item Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID}) [contributor]
  \item Colin Rundel \email{cr173@duke.edu} [contributor]
}

}
\keyword{internal}
