% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_tr.R
\docType{data}
\name{stopwords_tr}
\alias{stopwords_tr}
\title{A dataset of Turkish stopwords}
\format{
A tbl_df with with 92 rows and 1 variable:
\describe{
\item{word}{Stopword, in Turkish.}
}
}
\usage{
stopwords_tr
}
\description{
A dataset of stopwords used in Turkish text analysis.
}
\examples{
stopwords_tr
}
\keyword{datasets}
